+++ core/src/main/kotlin/wtf/nucker/kitpvpplus/config/LangConfig.kt
@@ -3,8 +3,8 @@ package wtf.nucker.kitpvpplus.config
 import net.kyori.adventure.text.Component
 import net.kyori.adventure.text.format.NamedTextColor
 import org.spongepowered.configurate.objectmapping.ConfigSerializable
-import wtf.nucker.kitpvpplus.statistics.DeathLang
-import wtf.nucker.kitpvpplus.statistics.StatisticsLang
+import wtf.nucker.kitpvpplus.statistics.lang.DeathLang
+import wtf.nucker.kitpvpplus.statistics.lang.StatisticsLang
 import wtf.nucker.kitpvpplus.util.KotlinExtensions.component
 
 @ConfigSerializable
