+++ core/src/main/kotlin/wtf/nucker/kitpvpplus/config/SettingsConfig.kt
@@ -2,7 +2,7 @@ package wtf.nucker.kitpvpplus.config
 
 import org.spongepowered.configurate.objectmapping.ConfigSerializable
 import wtf.nucker.kitpvpplus.database.DatabaseConfigSettings
-import wtf.nucker.kitpvpplus.statistics.EconomySettings
+import wtf.nucker.kitpvpplus.statistics.config.EconomySettings
 
 @ConfigSerializable
 class SettingsConfig {
